requirements.txt
fastapi 
uvicorn 
pypdf2 
langchain 
google-generativeai 
jinja2
faiss